Define S2KCountMin S2KCountMax S2KCountCommon
These are the magic iteration count numbers:

* S2KCountMin     = 1024
* S2KCountDefault = 65536      // by convention
* S2KCountMax     = 65011712

Define S2kCountDefault to make it explicit what iteration count is used
if none is specified.

(Aside, it's not clear why this 65535 value is used: it's a
significantly worse default than GnuPG 2.1.x which defaults to
26214400)

By naming this value as a default it should be clearer how to ratchet up
the iteration count in the future.

See RFC 4880 https://tools.ietf.org/html/rfc4880#section-3.7.1.3