openpgp/packet/signature: output reason for revocation
When building subpackets for a Signature, ensure a "reason for
revocation" subpacket is added if the Signature is a
SigTypeKeyRevocation.

Previously, the Signature fields `RevocationReason` and
`RevocationReasonText` were used only when *parsing* a revocation. This
change means they're also output in a subpacket if a revocation
signature is generated.

See https://tools.ietf.org/html/rfc4880#section-5.2.3.23