sshd/terminal: Add Terminal.SetEnterClear(bool)
SetEnterClear controls whether to keep the input line after enter by
pushing \r\n, or to clear the line into a fresh prompt without echo'ing
it above the prompt.

Using SetEnterClear allows us to control what content is rendered on
enter, so it does not have to be a verbatim echo.

For example, ssh-chat uses this to render timestamps in front of each
line.

The default behaviour continues to be backwards-compatible.