Add support for AEAD Encrypted Data Packets (#36)
And V5 Symmetrically Encrypted Data Packets (which use AEAD as well).

To enable AEAD when encrypting messages symmetrically, pass:

    &packet.Config{
        AEADConfig: &packet.AEADConfig{}
    }

to `openpgp.EncryptSymmetrically`. To enable AEAD when using public-key
cryptography, pass the same config to `openpgp.NewEntity` to announce
support for AEAD in the newly generated key. This library will
automatically use AEAD when encrypting to keys that claim to support it.

It is also possible to change the AEAD mode that is used when encrypting
symmetrically, or the preferred AEAD mode when generating a key, by
passing:

        AEADConfig: &packet.AEADConfig{
            DefaultMode: packet.AEADModeEAX
        }

`AEADModeEAX` is the default. The other options are `AEADModeOCB` and
`AEADModeExperimentalGCM`. The latter is a private/experimental
algorithm, and not defined in the spec. It is also supported by
OpenPGP.js, but not, at the moment, GPG. It is typically the most
performant option, as GCM is implemented in asm.

And finally, when encrypting (either symmetrically or asymmetrically),
it is also possible to pass the chunk size that is used:

            ChunkSize: 1 << 10 // 1KiB

A smaller chunk size means that, when en/decrypting a stream, you'll
receive data sooner (and more often). It also means a marginally bigger
overhead (16 bytes per chunk). The default is 256KiB, which we feel
strikes a decent balance, although that might not be the case for
gigantic messages. The passed chunk size is rounded down to the closest
power of two and clamped between 64 bytes and 128 MiB.
(According to the spec, chunk sizes larger than 128 MiB (1 << 27 bytes)
SHOULD NOT be used, although the packet format allows it.)

Note: when using SerializeSymmetricKeyEncryptedReuseKey, one must now
set a non-nil config.AEADConfig when used with an AEAD Encrypted Data
Packet, and similarly one must call SerializeAEADEncrypted after calling
SerializeSymmetricKeyEncryptedReuseKey with a non-nil config.AEADConfig.

See https://tools.ietf.org/html/draft-ietf-openpgp-rfc4880bis-08.