ssh: add support for server side multi-step authentication
- add ErrPartialSuccess. Authentication callbacks must return this error for
multi-step authentication when a specific authentication step succeed
- add PartialSuccessMethods to ConnMetadata interface, it returns the ordered
list of authentication methods that returned ErrPartialSuccess. It can be used
inside callbacks to find if a multi-step authentication is done using the
correct sequence and to return the authentication methods that can continue
- add NextAuthMethodsCallback, this callback is called when an authentication
callback returns ErrPartialSuccess or if, after an initial partial success, an
authentication step fails. It must return the list of authentications methods
that can continue.

This way an application can define per-user multi-step authentication.

Fixes #17889