crypto/ssh/terminal: add Terminal.GetPrompt method
The Terminal struct has no existing way of reading the value of the current
prompt. When updating the prompt it's sometimes necessary to know what the
existing value is. For example, when something needs to be appended/removed from
the existing prompt. `Terminal.prompt` is a private field and is set using
`SetPrompt`. This commit also includes a relevant test.