openpgp/packet: fix serialisation order of a RSA private key
This fix makes the serialisation/deserialisation round-trip produce
keys that stay equals, according to rsa.PrivateKey.Equal

As far as I understand this serialization problem isn't a security issue
as illustrated by the test that perform this round-trip then sign some data.
This means that the keys were equivalent, but not striclty equal.