ssh: implement partial success in server
Add support for sending back partial success to the client while
handling authentication in the server. This is implemented by a special
error that can be returned by any of the authentication methods, which
contains the authentication methods to offer next. This way a server
can be implemented that e.g. only accepts public keys for the initial
authentication, and only after partial success presents a keyboard
interactive authentication method.

Signed-off-by: Peter Verraedt <peter.verraedt@kuleuven.be>