ssh/knownhosts: check more than one key
The problem was that it was keeping only the first known key of each
type found. If you have a server advertising multiple keys of the same
type, you might get a missmatch key error.

Per sshd(8) man page, it should allow reapeatable hosts with different
host keys, although it don't specify anything about hosts being from
different types:

"It is permissible (but not recommended) to have several lines or
different host keys for the same names. This will inevitably happen when
short forms of host names from different domains are put in the file. It
is possible that the files contain conflicting information;
authentication is accepted if valid information can be found from either
file."

So, this changes knownhosts behavior to accept any of the keys for a
given host, regardless of type.

Fixes #36126

Signed-off-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>