ssh/agent: parse public key with ssh.ParsePublicKey
When the server gets an `agentSignRequest` it can be `ssh.PublicKey` or
`ssh.Certificate` formatted keys. As the server only parsed these into
`ssh.Key` they would require a `Marshal()` to `ParsePublicKey()` dance
when the passed key is a certificate.

Use `ssh.ParsePublicKey` instead of the `ssh.Key` struct directly.

Signed-off-by: Morten Linderud <morten@linderud.pw>